svelte(phase 3): delete session.js — relocate the live runtime into SessionPage
The remaining runSessionApp head moves into <SessionPage>'s onMount as a local
startSessionRuntime(): per-page settings/markdown bootstrap, setupSessionUi, the
content-runtime wiring, the mobile-layout/close-sidebar bridges, header-handler
binding, the initial navigateTo, and the annotation-layer init.

Two focused utils carry the reusable pieces: lazy-highlight.js
(applyLazyHighlighting) and session-content-runtime.js (wireSessionContentRuntime
— builds the entry renderer + sessionFormat, assigns contentRuntime.renderEntry/
afterRender, and binds the single delegated copy/fork/label handler on #messages).

session/session.js + session.test.js are removed (the dead re-exports went with
them; reconcile is unit-tested on the model). The Go toggle-state guard test is
repointed from session.js to session-content-runtime.js.
