+++ docs/dev/svelte-migration-plan.md
@@ -601,6 +601,9 @@ per increment.
 | ✅ DONE + verified | **Artifacts collection → `<ArtifactPanel>` (reactive).** The panel reads the shared `SessionDataModel` via context and recomputes the visible artifact set (+ tab count + enabled-state fallback) through a `$derived` feeding an `untrack`'d sync `$effect`, replacing session.js's imperative `refreshArtifacts` + the `window.__piArtifactPanel.setArtifacts` push. The help (?) modal wiring moved into `<RightSidebar>`. `session.js` lost `refreshArtifacts`/`applyArtifactsEnabled`/the artifact-host block + the `collectArtifacts`/`filterArtifacts`/`readArtifactSettings` imports. **Bug caught by e2e:** the first cut put collection in an `$effect` calling `setArtifacts` (which reads/writes `artifacts`/`selectedId` `$state`) → `effect_update_depth_exceeded`; fixed by the `$derived` + `untrack` split. Verified: web 528 + knip clean + no a11y warnings + `go test ./internal/ui` + Desktop-Chrome e2e 56/2 (+1 documented load-earlier retry). |
 | ✅ DONE + verified | **Model reconciliation → `SessionDataModel.reconcile()`.** Moved session.js's `syncDataModelEntries`/`replaceMapContents` (live-reload + load-earlier path) onto the model as `reconcile(entries)`: in-place entries splice + lookup-map refills + advance the active leaf to the newest descendant. `session.js` now just delegates (`dataModel.reconcile?.(entries)`), exposed on `window.__piReconcileEntries` for `<LiveReload>` + load-earlier; dropped the `buildSessionLookups`/`buildTree`/`buildTreeNodeMap`/`findNewestLeaf` body imports. `session.test.js` updated to provide the reactive model (mirroring SessionPage); 3 reconcile unit tests added. Verified: web 531 + knip clean + build + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (live-reload/load-earlier/chat green). |
 | ✅ DONE + verified | **Page-global glue → `session-globals.js` (called by SessionPage).** Relocated the "tail" of `runSessionApp` — done-notifier, `setupKeyboardNav`, version checker, session-list palette (+ Cmd+K), the global keyboard shortcuts (Cmd+K/B/T/⇧L/⇧N//), the shortcuts/new-session relay buttons, the load-earlier banner, and the mobile visual-viewport/scroll-lock handlers — into `setupSessionGlobals({ windowImpl, documentImpl, model, sessionId, navigateTo })`. It tracks the listeners it adds and returns a disposer that `<SessionPage>` calls on unmount. `window.__piReconcileEntries` now set by `<SessionPage>` (from `model.reconcile`) *before* the child components mount, closing the reload race. `session.js` dropped 6 imports + ~175 lines (now 296). Added `session-globals.test.js` covering the keyboard shortcuts (no e2e for those). Verified: web 538 + knip clean + build (no a11y) + `go test ./internal/ui` + go vet + Desktop-Chrome e2e 56/2 (+1 documented load-earlier retry). |
+| ✅ DONE + verified | **`session.js` DELETED — head relocated into `<SessionPage>` + two utils.** The remaining `runSessionApp` head moved to `<SessionPage>`'s `onMount` as a local `startSessionRuntime()`: per-page settings/markdown bootstrap, `setupSessionUi`, the content-runtime wiring, the `__piIsMobileLayout`/`__piCloseSidebar` bridges, `attachHeaderHandlers`, the initial `navigateTo`, and the annotation-layer `init({...})`. Two extracted utils: **`lazy-highlight.js`** (`applyLazyHighlighting`) and **`session-content-runtime.js`** (`wireSessionContentRuntime`: builds the entry renderer + `sessionFormat`, assigns `contentRuntime.renderEntry`/`afterRender`, sets `downloadSessionJson`, and binds the single delegated copy/fork/label handler on `#messages`). **`session/session.js` + `session.test.js` DELETED** (the dead re-exports went with it; reconcile is unit-tested on the model). Go guard `TestNavigationReappliesCurrentToggleStateAfterRenderingMessages` repointed from `session.js` → `session-content-runtime.js`. Verified: web 535 + knip clean + build (no a11y) + `go test ./internal/ui` + go vet + **Desktop-Chrome e2e 57/2 (clean, no flake)**. |
+
+**`session.js` teardown COMPLETE** — the live session runtime is now `<SessionPage>` (orchestration) + `session-globals.js` + `session-content-runtime.js` + `lazy-highlight.js`, all reading the reactive `SessionDataModel`. The remaining §8 deletions (the chat/live runner modules still owned by `<ChatComposer>`/`<LiveReload>`, and decomposing `session-entry-renderer.js` into Svelte sub-components) are the next tranche.
 | 🔶 remaining | **Phase 2 was NOT isolatable from `session.js` (resolved for tree+header).** The tree's rendering, the active leaf/target (`currentLeafId`/`currentTargetId`), and `filterMode`/`searchQuery` all live as imperative locals in `session.js` (and `export-entry.js`); tree clicks drive **content** rendering via the navigator. Swapping only the tree DOM to `<SessionTreeNodes>` would need a throwaway bridge between that imperative state and the reactive model — which Phase 3 then deletes. **Do the tree cut-over together with moving navigation + filter state into `SessionDataModel`** (i.e. merge the front of Phase 3 into Phase 2), so the model is the single source of truth and `session.js`/`export-entry.js` stop owning that state. The staged `TreeNode`/`SessionTreeNodes` components are ready for that step. |
 
 **Recommended next step (combined Phase 2/3a):** move `currentLeafId`,
